Add delete car functionality
A new function `delete_car` is added to the views to allow users to delete an existing car from the database. Also, the corresponding URL route is added to urls.py. The delete button is added in the get_car_by_id template while the placeholder text in the search bar in the cars and get_car_by_id templates is changed to "Поиск".