Add functionality to add a new car
Added a new 'add_car' view in views.py which handles both GET and POST requests to add a new car to the database. Adjusting the routing to include a path for the new function in urls.py. Created a new 'add_car.html' template for the front-end presentation and data entry form. Additionally, updated navigation and adjusted form action paths on 'get_car_by_id.html' as well as 'cars.html'.